refactor: UserAgent 성격 관련 프롬프트 메소드 위치 변경
